+++ src/Presenter/Presenter.java
@@ -1,6 +1,7 @@
 package Presenter;
 
 import model.Service;
+import model.human.Gender;
 import view.View;
 
 public class Presenter {
@@ -12,12 +13,12 @@ public class Presenter {
         this.service = new Service();
     }
 
-    public void getFamilyTreeInfo() {
-        this.service.getFamilyTreeInfo();
+    public String getFamilyTreeInfo() {
+       return this.service.getFamilyTreeInfo();
     }
 
-    public void addNewHumanToFamilyTree() {
-        this.service.addNewHumanToFamilyTree();
+    public void addNewHumanToFamilyTree(Gender gender, String name, String lastName, Integer age ) {
+        this.service.addNewHumanToFamilyTree(gender, name, lastName, age);
     }
 
     public void writeFamilyTreeInFile(String path) {
