+++ src/model/familytree/FamilyTreeItem.java
@@ -8,4 +8,5 @@ public interface FamilyTreeItem {
     String getLastname();
     Gender getGender();
     Integer getId();
+    String getGeneralInfo();
 }
