+++ src/view/View.java
@@ -7,5 +7,4 @@ public interface View {
     void start();
     void changeFlag();
     Presenter getPresenter();
-
 }
